src/test/java/uoa/lavs/mainframe/simulator/nitrite/LoadLoanPaymentsTests.java
package uoa.lavs.mainframe.simulator.nitrite;

import org.junit.jupiter.api.Test;
import uoa.lavs.mainframe.Connection;
import uoa.lavs.mainframe.Status;
import uoa.lavs.mainframe.messages.loan.LoadLoanPayments;
import uoa.lavs.mainframe.simulator.NitriteConnection;

import static org.junit.jupiter.api.Assertions.*;
import static uoa.lavs.mainframe.MessageErrorStatus.LOAN_NOT_FOUND;

class LoadLoanPaymentsTests {
    @Test
    public void handlesMissingLoan() {
        // Arrange
        Connection connection = new NitriteConnection();
        LoadLoanPayments message = new LoadLoanPayments();
        message.setLoanId("123-08");

        // Act
        Status status = message.send(connection);

        // Assert
        assertAll(
                () -> assertEquals(LOAN_NOT_FOUND.getCode(), status.getErrorCode()),
                () -> assertEquals(LOAN_NOT_FOUND.getMessage(), status.getErrorMessage())
        );
    }

    @Test
    public void handlesExistingLoan() {
        // Arrange
        Connection connection = new NitriteConnection(
                DatabaseHelper.generateDefaultDatabase());
        LoadLoanPayments message = new LoadLoanPayments();
        message.setLoanId("123-09");

        // Act
        Status status = message.send(connection);

        // Assert
        assertAll(
                () -> assertTrue(status.getWasSuccessful()),
                () -> assertEquals("123", message.getCustomerIdFromServer()),
                () -> assertEquals("John Doe", message.getCustomerNameFromServer()),
                () -> assertEquals(2, message.getPageCountFromServer()),
                () -> assertEquals(19, message.getPaymentCountFromServer()),
                () -> assertEquals(63.75, message.getPaymentInterestFromServer(1)),
                () -> assertEquals(1, message.getPaymentNumberFromServer(1)),
                () -> assertEquals(509.25, message.getPaymentPrincipalFromServer(1)),
                () -> assertEquals(9490.75, message.getPaymentRemainingFromServer(1))
        );
    }
}